Sua loi

+++ QuanLyQuanCafe/fThongTinTaiKhoan.cs
@@ -28,10 +28,11 @@ namespace QuanLyQuanCafe
             loginAccount = acc;
             LoadAccountList();
         }
+        public fThongTinTaiKhoan() { }
         #region Account
         void ChangeAccount(Account acc)
         {
-            
+
         }
         #endregion
         BindingSource bindingAccount = new BindingSource();
@@ -44,7 +45,8 @@ namespace QuanLyQuanCafe
         void LoadAccount()
         {
             bindingAccount.DataSource = AccountDAO.Instance.Account();
-            dtgvAccount.DataSource = bindingAccount;
+            try { dtgvAccount.DataSource = bindingAccount; }
+            catch { }
         }
 
         private void btnXem_Click(object sender, EventArgs e)
@@ -66,36 +68,81 @@ namespace QuanLyQuanCafe
             string tenHienThi = txbHienThi.Text;
             string pass = txbPass.Text;
             int type = (int)nmType.Value;
-            
-            int id = Convert.ToInt32(txbID.Text);
-
-            Account accountTrung = AccountDAO.Instance.AccountByUser(tenTaiKhoan);
-
-            if (accountTrung.IdAccount != id && accountTrung.TenTaiKhoan == tenTaiKhoan)
-                MessageBox.Show("Tên muốn sửa đã có");
+            if (String.IsNullOrEmpty(txbID.Text))
+                MessageBox.Show("Không có gì để thay đổi");
             else
             {
-                if (AccountDAO.Instance.UpdateListAccount(id, tenTaiKhoan, tenHienThi, pass, type))
+                int id = Convert.ToInt32(txbID.Text);
+                updateAccInDatabase(id, tenTaiKhoan, tenHienThi, pass, type);
+            }
+        }
+        public bool updateAccInDatabase(int id, string tenTaiKhoan, string tenHienThi, string pass, int type)
+        {
+            if (tenTaiKhoan == "" || tenHienThi == "" || id <= -1)
+            {
+                MessageBox.Show("Vui lòng nhập đầy đủ ! Không được để trống ");
+                return false;
+            }
+            else
+            {
+                Account accountTrung = AccountDAO.Instance.AccountByUser(tenTaiKhoan);
+
+                if (accountTrung != null && accountTrung.IdAccount == id)
                 {
-                    MessageBox.Show("Sửa thành công");
-                    LoadAccount();
+                    if (AccountDAO.Instance.UpdateListAccount(id, accountTrung.TenTaiKhoan, tenHienThi, pass, type))
+                    {
+                        MessageBox.Show("Sửa thành công");
+                        LoadAccount();
+                        return true;
+                    }
+                    else
+                    {
+                        MessageBox.Show("Sửa không thành công");
+                        return false;
+                    }
+                }
+                if (accountTrung != null)
+                {
+                    MessageBox.Show("Tên muốn sửa đã có");
+                    return false;
                 }
                 else
-                    MessageBox.Show("Sửa không thành công");
+                {
+                    if (AccountDAO.Instance.UpdateListAccount(id, tenTaiKhoan, tenHienThi, pass, type))
+                    {
+                        MessageBox.Show("Sửa thành công");
+                        LoadAccount();
+                        return true;
+                    }
+                    else
+                    {
+                        MessageBox.Show("Sửa không thành công");
+                        return false;
+                    }
+                }
             }
         }
         private void btnXoa_Click(object sender, EventArgs e)
         {
-            int id = Convert.ToInt32(txbID.Text);
-            DeleteAccount(id);
-        }
-        void DeleteAccount(int id)
-        {
-            if(loginAccount.IdAccount.Equals(id))
+            if (String.IsNullOrEmpty(txbID.Text))
+                MessageBox.Show("Không có gì để thay đổi");
+            else
             {
-                MessageBox.Show("Không thể xóa tài khoản đang đăng nhập");
-                return;
+                int id = Convert.ToInt32(txbID.Text);
+                DeleteAccount(id);
             }
+        }
+        public void DeleteAccount(int id)
+        {
+            try
+            {
+                if (loginAccount.IdAccount.Equals(id))
+                {
+                    MessageBox.Show("Không thể xóa tài khoản đang đăng nhập");
+                    return;
+                }
+            }
+            catch { }
             if (AccountDAO.Instance.DeleteAccount(id))
             {
                 MessageBox.Show("Xóa thành công");
@@ -110,20 +157,46 @@ namespace QuanLyQuanCafe
             string tenTaiKhoan = txbTen.Text;
             string tenHienThi = txbHienThi.Text;
             int type = (int)nmType.Value;
-            if (AccountDAO.Instance.AccountByUser(tenTaiKhoan) == null)
-            {
-                if (AccountDAO.Instance.InsertAccount(tenTaiKhoan, tenHienThi, type))
-                {
-                    MessageBox.Show("Thêm thành công");
-                    LoadAccount();
-                }
-                else
-                    MessageBox.Show("Thêm không thành công");
-            }
-            else
-                MessageBox.Show("Đã tồn tại tên tài khoản " + tenTaiKhoan);
+
+            AddAccToDatabase(tenTaiKhoan, tenHienThi, type);
         }
 
+        public bool AddAccToDatabase(string tenTaiKhoan, string tenHienThi, int type)
+        {
+            if (tenTaiKhoan == "" || tenHienThi == "") // sửa lỗi nhập giá trị null
+            {
+                MessageBox.Show("Vui lòng nhập đầy đủ ! Không được để trống ");
+                return false;
+            }
+            else
+            {
+                if (AccountDAO.Instance.AccountByUser(tenTaiKhoan) == null)
+                {
+                    if (AccountDAO.Instance.InsertAccount(tenTaiKhoan, tenHienThi, type))
+                    {
+                        MessageBox.Show("Thêm thành công");
+                        LoadAccount();
+                        return true;
+                    }
+                    else
+                    {
+                        MessageBox.Show("Thêm không thành công");
+                        return false;
+                    }
+                }
+                else
+                {
+                    MessageBox.Show("Đã tồn tại tên tài khoản " + tenTaiKhoan);
+                    return false;
+                }
+
+            }
+        }
+        public int countAccount()
+        {
+            int a = AccountDAO.Instance.countAcc();
+            return a;
+        }
         private void button1_Click(object sender, EventArgs e)
         {
             txbHienThi.Clear();
